pyproject.toml
[project]
name = "opus"
version = "0.1.0"
description = "Terminal-based AI agent for software engineering automation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "google-genai>=0.3.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "httpx>=0.27.0",
    "html2text>=2020.1.16",
]

[project.scripts]
opus = "opus.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/opus"]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
